+++ src/renderer/index.js
@@ -17,8 +17,8 @@ import {
   doOpenModal,
   doHideModal,
 } from 'redux/actions/app';
-import { doToast, doBlackListedOutpointsSubscribe, isURIValid } from 'lbry-redux';
-import { doNavigate } from 'redux/actions/navigation';
+import { doToast, doBlackListedOutpointsSubscribe, isURIValid, setSearchApi } from 'lbry-redux';
+import { doNavigate, doHistoryBack, doHistoryForward } from 'redux/actions/navigation';
 import { doDownloadLanguages, doUpdateIsNightAsync } from 'redux/actions/settings';
 import { doAuthenticate, Lbryio, rewards } from 'lbryinc';
 import 'scss/all.scss';
@@ -36,6 +36,18 @@ autoUpdater.logger = remote.require('electron-log');
 if (process.env.LBRY_API_URL) {
   Lbryio.setLocalApi(process.env.LBRY_API_URL);
 }
+
+if (process.env.SEARCH_API_URL) {
+  setSearchApi(process.env.SEARCH_API_URL);
+}
+
+ipcRenderer.on('navigate-backward', () => {
+  app.store.dispatch(doHistoryBack());
+});
+
+ipcRenderer.on('navigate-forward', () => {
+  app.store.dispatch(doHistoryForward());
+});
 
 // We need to override Lbryio for getting/setting the authToken
 // We interect with ipcRenderer to get the auth key from a users keyring