+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Dark theme as default ([#2210](https://github.com/lbryio/lbry-desktop/pull/2210))
+- Invite page improvements including increase of reward to 20 LBC / 10 redemptions ([#2256](https://github.com/lbryio/lbry-desktop/pull/2256))
 - Less intrusive first run flow, email collection ([#2210](https://github.com/lbryio/lbry-desktop/pull/2210))
 - Add icon for your own publishes on cards/tiles ([#2249](https://github.com/lbryio/lbry-desktop/pull/2249))
 - Improved Claimed Reward messaging ([#2253](https://github.com/lbryio/lbry-desktop/pull/2253))